Remove Web and add Mac
I just commented out the code for the styler and got the program to run. Since I'm on a mac, I had to change the way it found the directory for courses since mac uses '/' instead of '\\'. idk if that will work for windows so I just commented it out